twitter_clone/pyproject.toml
[tool.poetry]
name = "twitter-clone"
version = "0.1.0"
description = "Twitter Clone"
authors = ["DiegoSanchez <diegosmtattoo@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
django = ">=5.1.6,<6.0.0"
djangorestframework = ">=3.15.2,<4.0.0"
dj-database-url = ">=2.3.0,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
djangorestframework-simplejwt = ">=5.4.0,<6.0.0"
django-cors-headers = ">=4.7.0,<5.0.0"
gunicorn = ">=23.0.0,<24.0.0"
whitenoise = ">=5.3.0,<6.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
